Implement incremental backoff and throttle, log at end
to prevent hammering the site, implemented an incremental backoff in the scenario where a page doesn't load.

Maximum speed of 1 request per second so that it doesn't hammer the site too hard.

Writes the links visited (and the 3rd level links not) into a json at the end for further inspection and analysis.